Trait sp_consensus_aura::digests::CompatibleDigestItem

pub trait CompatibleDigestItem<Signature>: Sized {
    fn aura_seal(signature: Signature) -> Self;
    fn as_aura_seal(&self) -> Option<Signature>;
    fn aura_pre_digest(slot: Slot) -> Self;
    fn as_aura_pre_digest(&self) -> Option<Slot>;
}

A digest item which is usable with aura consensus.

Required Methods

fn aura_seal(signature: Signature) -> Self

Construct a digest item which contains a signature on the hash.

fn as_aura_seal(&self) -> Option<Signature>

If this item is an Aura seal, return the signature.

fn aura_pre_digest(slot: Slot) -> Self

Construct a digest item which contains the slot number

fn as_aura_pre_digest(&self) -> Option<Slot>

If this item is an AuRa pre-digest, return the slot number